+++ monitor-main/src/main/kotlin/com/jtmnetwork/monitor/core/domain/model/Event.kt
@@ -1,3 +1,3 @@
-package com.jtmnetwork.main.core.domain.model
+package com.jtmnetwork.monitor.core.domain.model
 
 data class Event(val name: String, val value: String)
