+++ monitor-main/src/main/kotlin/com/jtmnetwork/monitor/core/usecase/handler/EventHandler.kt
@@ -1,6 +1,6 @@
-package com.jtmnetwork.main.core.usecase.handler
+package com.jtmnetwork.monitor.core.usecase.handler
 
-import com.jtmnetwork.main.core.domain.model.Event
+import com.jtmnetwork.monitor.core.domain.model.Event
 import okhttp3.WebSocket
 
 interface EventHandler {
