+++ monitor-spigot/build.gradle
@@ -49,7 +49,7 @@ dependencies {
     implementation "com.jtm.framework:framework:0.9.11"
     implementation "com.fasterxml.jackson.core:jackson-databind:2.12.5"
     implementation "com.fasterxml.jackson.module:jackson-module-kotlin:2.12.5"
-    implementation "com.jtmnetwork.main:plugin:0.15"
+    implementation "com.jtmnetwork.main:plugin:0.16"
 
     compileClasspath 'me.clip:placeholderapi:2.10.9'
